[NFC] Move InsertOrdered{Set,Map} into a new header
Move the `InsertOrderedSet` and `InsertOrderedMap` implementations out of
Relooper.h and into a new insert_ordered.h so they can be used more widely. Does
not change any of the implementation code.
